Message Editor

+++ src/main/java/org/example/repos/UserRepo.java
@@ -3,7 +3,7 @@ package org.example.repos;
 import org.example.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User,Long> {
+public interface UserRepo extends JpaRepository<User,Integer> {
     User findByUsername(String username);
 
     User findByActivationCode(String code);
